+++ src/parsers/index.ts
@@ -11,6 +11,7 @@ export { OpenApiParser } from "./openapi/OpenApiParser.js";
 export { SwaggerParser } from "./swagger/SwaggerParser.js";
 export { GraphQLParser } from "./graphql/GraphQLParser.js";
 export { DirectoryParser } from "./graphql/DirectoryParser.js";
+export { GrpcParser } from "./grpc/GrpcParser.js";
 
 // Utilities
 export { detectFormat } from "./utils/auto-detector.js";